Bulk change (and adds no-littering package)
Primarily fixes an issue with the config that was due to an extra
parens, but then I am not sure why the diff is such a mess.

Learnings: prefer smaller changes, and once the diff is not simple,
then try figure what is the issue causing these crazy diffs.

Also added the `no-littering` package.