Remove unneeded syntax entry for elisp
The entry was added originally to be able to operate on elisp
identifiers using '/' and '-' characters, but this changes comes at a
cost of not being able to operate on the parts of the words. The
solution is to use the 'o' text-object which will match the full
identifier e.g. `cio`, `vio`, etc.