Disable package GCMH
I am not sure if I was using the package correctly, but I was running
into an issue where saving files was taking a long time.  I noticed
that the value of `gc-cons-threshold` was super high, and so I think
the max value that was being set in early-init.el was not actually
getting reduced afterwards.  I also read in a Reddit post that GCMH
package is not very efficient with its use of timers.  So we are going
to simply go with the default value.