Fix broken evil-snipe
Honestly I am not sure why it was not working before. I tried several
things, and putting the `(evil-snipe-mode +1)` in the `:init` section
got things working, but ChatGPT said that this was necessary, and that
makes sense, and then moving it back to `:config` and testing,
everything was working again.  :shrug: