Remove custom project compile/test commands
We don't need to create a different set of variables for project compile/test commands.  We will instead use the
in-built `project-compile` command, which itself uses the in-built `compile-command` variable (but using the project
root directory as working directory). Now of course a project often has many more commands than just a compile command
e.g. format, lint, test, etc. So instead of trying to special case a Test command, we will rely on the projects build
runner for executing the various sub-commands, by simply using arguments to the `compile-command`.  For large projects
we can use a completing-read interface to the runner e.g. make, just, gradle, etc.