Add prog-mode and LSP bindings to localleader
Instead of having global bindings that don't really apply to all
modes, we add the bindings to the mode map, which is accesible via our
localleader/mode binding map e.g. "SPC m".

We also add discoverability for auto scaling up functionality
e.g. instead of enabling LSP by default (which can be heavy for large
projects), we provide bindings for discoverability, which indicate
that LSP should be started to use it.  Then when LSP is enabled it
shadows those bindings "upgrading" them.