Adds comments on usage of consult-imenu
We went on long adventure trying to improve consult-imenu
classification of lang items, but enabling LSP mode results in a
different index of items being fed to imenu, so ultimately we had to
give up. We add the use-package section as a reminder incase we forget
and attempt this adventure again at some point.